REFACTOR test classes into `Assets`, `Unit` & `Feature` namespaces
